+++ src/com/amazon/gamelift/agent/model/exception/NotFoundException.java
@@ -12,13 +12,13 @@ public class NotFoundException extends AgentException {
      * Throws NotFoundException with message and throwable
      */
     public NotFoundException(final String message, final Throwable exception) {
-        super(message, exception, false);
+        super(message, exception, true);
     }
 
     /**
      * Throws NotFoundException with message
      */
     public NotFoundException(final String message) {
-        super(message, false);
+        super(message, true);
     }
 }