+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 CHANGELOG
 
+# 1.0.3 (12/30/2024)
+- Add termination reason to NotifyServerProcessTermination message
+- Enable compute registration by default when running on non-container compute
+- Minor bug fixes
+
 # 1.0.2 (11/04/2024)
 - Update compute registration for container fleets
 - Skip spot termination checks when running on non-spot or non-EC2 instances